Implemented ticket #375 "Create an installer or package creator"
git-svn-id: https://svn.code.sf.net/p/numere/source/trunk@542 d50cd26f-24d5-421e-9e7d-30b8b4216d91

+++ gui/dialogs/debugviewer.cpp
@@ -239,7 +239,7 @@ void DebugViewer::getInformationByStackId(size_t id)
     // trace list control
     for (int i = 0; i < m_stacktrace->GetItemCount(); i++)
     {
-        if (m_stacktrace->GetItemCount() - 1 - i == id)
+        if (m_stacktrace->GetItemCount() - 1 - i == (int)id)
             m_stacktrace->SetItemText(i, "->");
         else
             m_stacktrace->SetItemText(i, "");
